Take the configuration from the environment
This is a way to keep the app stateless in accordance with the 12-factor
apps principles. There are both required variables, that will throw a
KeyError if not given (like SECRET_KEY) and optional variables with
defaults.

If we run the app manually, we need to specify the config inline, which
is not ideal. The solution would be using dotenv, but as in this example
we run everything with docker, the docker-compose.yml will provide all
the configuration in development.